created new layout for static iran view
this layout contains, inside of it and with some modifications (as not
json, took out \" for instance) all of the data from the iran test data
js file. server client contains the comments, and the index.html file
contains the articles. note that this layout is still dependent on the
bootstrap js files for css styling, but the index file also contains
styles in the head that modify the bootstrap.css file.