changes on desktop version of prototype
there are continuing errors that have been introduced by the flip thing I am trying to debug. specifically, when the flip classes are added it interferes with some of the normal bootstrap functioning. Collapsing no longer works as expected as things get covered up, and the links open in a new tab.