Using lawnchair to store comments on client
NOTE: I removed the window-name adaptor
The window-name adaptor was wrongly configured as default.
If we keep using lawnchair, we should be sure to get a proper build of
it.

These comment records get marked with the users
reaction, which is used to determine which comments
we've shown.

Using Lawnchair since it's nicer than straight localStorage,
not using indexedDB or websql yet, since that gets into
prompting the user for permission, etc, depending on
how much storage you request.  requires a bit of research.