threaded poller
Use standard java wait and notify to broadcast to worker threads
when there are new votes to process.

Ensure no two worker threads are working on the same zid by keeping
the in-progress zids in a set.

Track the votes which need processing in a map of lists for each zid.