add meta tag for CloudFlare SSL
Here's the email they sent:

You have chosen the manual SSL configuration option for pol.is. There is
one more step before your website can begin receiving the performance
and security benefits of CloudFlare. We’ll walk you through that step:
adding the meta tag to pol.is. Once it’s verified, your website will be
on CloudFlare and run faster, safer and smarter.

ADD THE META TAG TO pol.is:
STEP A. Please insert the following unique meta tag into the head
section on either pol.is or www.pol.is. It must be inserted immediately
after the opening <head> section. The unique <meta> tag for your website
is:

    <meta name="globalsign-domain-verification"
    content="Dp-UFZfqPtVw6GK7XwmqFPmgjnUyTmU7xNGwpMnr-K" />

    Please note that the meta tag is unique for this website. If you
    have a second website, you will need a second SSL meta tag. SSL
    cannot be verified manually using any other subdomains.

    STEP B. Confirm that CloudFlare has verified your SSL tag for the
    subdomains that you chose during the sign-up process by logging into
    your account (it may take a few hours):
    www.cloudflare.com/dns-settings?z=pol.is

    If SSL isn’t working on your website, please review these
    troubleshooting tips:
    support.cloudflare.com/hc/en-us/articles/200170566-My-SSL-isn-t-working-Why-not-

    If you need help, please visit www.cloudflare.com/support

    Thanks for joining the CloudFlare community!

    The CloudFlare Team

    P.S. If you want to use the automatic SSL process instead, please
    login to your CloudFlare account and un-pause the service. This will
    immediately restart the automatic SSL process, which requires no
    changes on your end. CloudFlare will automatically issue a SSL
    certificate that takes up to 10 minutes to activate. During this
    time, visitors may see a warning when they visit your website over
    https://. Once the SSL activates, the SSL warnings will disappear.