badwords
The idea for now is to use a basic badwords list, and to
set the comment's velocity to zero, so that it is not
sent out to other users.

A moderation email is sent (to us for now), with a signed URL
which can be used to easily unmute the comment in one click
from a cellphone.

If this starts to become a scaling problem, we can solve the issue then.