make all metadata required for now HACK
Should be stored in the DB.
Probably under on the question rows.
But here we're serving it up as part of the answers - why?
It makes it easier to render the answers with radio buttons.
It's a pain to plumb that data through to the views, because they're
dynamically created by Thorax collection views.