use connection pool for postgres
Here's what I looked at when making this choice:

I read this thread:
https://github.com/brianc/node-postgres/issues/227
And considered any-db
https://github.com/grncdr/node-any-db

But I didn't check if using any-db would get in the way of using
postgres-specific queries. If any-db were more popular, i might have
looked further.

I chose to go with the pooling that was added to node-postgres after the
discussion on issue 227.

node-postgres is more active, so this pooling code probably gets run a lot.
https://github.com/brianc/node-postgres