getNextComment rewrite
voting returns the next comment in the response.

This removes the need for caching the next comment on the client, which
is important since we don't want to cache comments that may be later
removed.
It's also important to show the next comment quickly, so it's not ok to
wait for a separate ajax call.