WIP sid-zid, param getters return promises
To prevent creation of custom non-composable middleware param getters,
the param getters now return promises.  This allows for async param fetching.

One drawback with the async param fetchers is that subsequent param
getters can't run in parallel. This could be fixed by creating a big
parallel getter runner, which would also resolve auth. This would need
to take "want" or "need" as params, or the getters might be composed
within an optional(getFoo) function.

sid-zid change is still in progress. So far, inbox and conversation
work.