filter UA in app, not polisHost, use js to test for cookies, not ajax call
The ajax call cookie test was something i tried, which didn't solve the
problem. Putting the app on a subdomain did solve the problem. Don't
want the latency of the extra ajax call, so using js.

Having the filter logic on polisHost may encourage people to make their own
copy of polisHost, also, it's slow to update it, because no cachebuster
path.