big cluster id fix
The client code was written expecting that cluster ids were implied by
the index in the array in the JSON response. This was wrong - the IDs
are specified within each cluster object, and do not necessarily
correspond to the index of that object in the array of cluster objects.
(the reason for this is to allow new clusters to form, and gives us the
option of doing nice animations to preserve continuity or trigger
events/animations when groups appear or disband)

Previously, the raw members lists where extracted, and passed to the
vis as a list of lists. clusters[gid] = [4,5,6] for exapmple. Now the
entire cluster objects are sent to the vis as a map from gid to cluster
object.