pass pre-filled comment text form url to form
This is needed for the social login scenario where the user has already
filled out a comment. For simplicity, it's better to reload the page
when the user changes identity, so passing their comment text to the
reloaded page is unfortunate, but probably better than the bug farm that
will result from trying to support changing identity without a reload.