best-agree
Normally, we'd use n-success to indicate the number of agrees
(or disagrees) that made the comment repful for agreement
(or disagreement)
But in cases where we only have repful-for-disagree comments
for a group, we do a hack where we force a comment that is
generally agreed upon by the group to the front. That scenario
is indicated by the presence of the "best-agree" flag; in that
case, we should use "n-agree" as the count.