Add lots of PCA tests
I am trying to get full _branch_ coverage of the function
`power-iteration`, but failing to. There must be some corner cases I am
not surfacing (and Cursor neither).
I will leave it aside for now, make a note that it needs fixed, and will
go to cover the entirely missing lines in the wrapped-pca function.