Remove nonsensical test
Comparing the first two iterations of Sklearn and our power iteration
do not make sense, because we do not necessarily have the same internal
algorithm.
It might come a time where we want to dive to that level of detail,
for example when we want to move to sklearn instead
of custom PCA, but not now.