Add instrumentation utility
DISCLAIMER: Code written entirely with Cursor! I do not pretend to speak
Clojure fluently enough (or at all ;P) to code that myself. I have
supervised Cursor, probing it to generate tests, pass them, reviewing
the code.

This instrumentation utility records all the calls to whichever
functions we care to, storing in JSON the arguments and the return
values.

This will be handy to record the behaviour of the code on real
conversations, then comparing the behaviour of the python port to ensure
1:1 matching.