Enhance result comparison and add speed-up tracking
Note: the timing comparisons at the moment are very unfair to clojure, because in my dev setup clojure runs in docker whereas (for now) python runs on the full host. Thus python has access to more cores if numpy is properly multicore-optimized, and its BLAS calls are also compiled for the actual hardware (Apple Silicon), which is probably not the case in a linux docker image running on my mac I believe. So, fair comparison will be once we have all running in Docker. And ideally on a machine whose silicon matches the docker image.

- Improve result comparison with more detailed type checking
- Add timing information for Python vs Clojure function execution
- Introduce performance summary with speedup calculations
- Add more informative discrepancy reporting