Fix bug in PCA that caused different results
Found the bug ! (With Claude Code's help)
The PCA code starts by "cleaning" the matrix with some replacement rules for NaN
and strings. Then it proceeds to compute the PCA on that cleaned up matrix.
Great, I've done the cleaning, and done it in-place for efficiency, since the
matrix is cleaned up first thing in the code and the unclean one therefore not
used.  Right ?
...
RIGHT ??
*It turns out*, hidden way below, the projection of the participants on the
*low-dimensional space is (intentionally) done *on the non-cleaned matrix* !!
*(TODO : I'll have to put my math thinking cap on understand exactly why it was
*coded like that...)

Adding "copy=True" in one built-in invocation solved it.

This version here also restored the loop-in-loop cleanup code. My next commit will clean it up.