Turn off off the Clojure sign-flip for PCA
Of course this breaks the comparison to the golden files, as those included
the sign flip. So we're also giving the option in the regression test to ignore
PCA sign-flips. That will be handy later when we work on improving the PCA implementation,
as various PCA implementations have various sign conventions.